requirements.txt
beautifulsoup4==4.5.1
click==6.6
cookies==2.2.1
Flask==0.11.1
funcsigs==1.0.2
geopy==1.11.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==2.0.0
pbr==1.10.0
python-forecastio==1.3.4
requests==2.10.0
responses==0.5.1
simplejson==3.8.2
six==1.10.0
Werkzeug==0.11.10
wikipedia==1.4.0
wit==4.1.0
wolframalpha==2.4
